day_23/requirements.txt
chainlit==1.2.0
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-text-splitters==0.3.0
langchain-openai==0.2.0
faiss-cpu==1.9.0.post1
python-dotenv==1.0.1
requests==2.32.3